+++ shiny/source/global/helpers.R
@@ -5,26 +5,42 @@ prep_for_shiny <<- function(data, reorder_periods = reorder_periods) {
   data <- data %>%
     idaifield_as_matrix() %>%
     as.data.frame() %>%
+    mutate(relation.liesWithinLayer = na_if(relation.liesWithinLayer, "NA")) %>%
+    mutate(relation.liesWithinLayer = ifelse(is.na(relation.liesWithinLayer),
+                                             relation.liesWithin,
+                                             relation.liesWithinLayer)) %>%
+    mutate(relation.liesWithinLayer = ifelse(is.na(relation.liesWithinLayer),
+                                             relation.isRecordedIn,
+                                             relation.liesWithinLayer)) %>%
     remove_na_cols() %>%
     type.convert(as.is = FALSE)
 
   tryCatch({
     data <- data %>%
       mutate(date = as.Date(as.character(date), format = "%d.%m.%Y"))
-  }, error = function(e) print(paste("prep_for_shiny(): ", e)))
+  }, error = function(e) message(paste("Caught problem in prep_for_shiny(): ", e)))
 
   tryCatch({
     data <- data %>%
       mutate(beginningDate = as.Date(as.character(beginningDate), format = "%d.%m.%Y"))
-  }, error = function(e) print(paste("prep_for_shiny(): ", e)))
+  }, error = function(e) message(paste("Caught problem in prep_for_shiny(): ", e)))
 
   tryCatch({
     data <- data %>%
       mutate(endDate = as.Date(as.character(endDate), format = "%d.%m.%Y"))
-  }, error = function(e) print(paste("prep_for_shiny(): ", e)))
+  }, error = function(e) message(paste("Caught problem in prep_for_shiny(): ", e)))
 
 
   if(reorder_periods) {
+    if (length(data$period) == 0) {
+      data$period <- NA
+    }
+    if (length(data$period.start) == 0) {
+      data$period.start <- NA
+    }
+    if (length(data$period.end) == 0) {
+      data$period.end <- NA
+    }
     data <- data %>%
       mutate_at(c("period", "period.end", "period.start"), as.character) %>%
       # fix value for periods that have been assigned multiple periods
@@ -87,27 +103,6 @@ uid_by_operation <<- function(filter_operation = "all",
 }
 
 
-make_layer_selector <<- function(data_all, inputId,
-                                label = "Choose one or many contexts") {
-  # creates the selector for layers present in data_all as a relation
-
-  # first get the unique values that will be displayed in the selectInput
-
-  selectable_layers <- data_all$relation.liesWithinLayer %>%
-    unique() %>%
-    as.character() %>%
-    sort()
-  # produce the selectInput object that will be displayed
-  pickerInput(inputId = inputId,
-              label = label,
-              choices = selectable_layers,
-              multiple = TRUE,
-              selected = selectable_layers,
-              options = list("actions-box" = TRUE,
-                             "live-search" = TRUE,
-                             "live-search-normalize" = TRUE,
-                             "live-search-placeholder" = "Search here..."))
-}
 
 make_period_selector <<- function(inputId,
                                  label = "Select objects by periods") {
@@ -138,6 +133,32 @@ mq_spinner <<- function(object) {
 }
 
 
+milquant_plotly_layout <<- function(plotly_fig, caption = FALSE) {
+  plotly_fig <- plotly_fig %>%
+    config(displaylogo = FALSE,
+           modeBarButtonsToRemove = c("select2d", "lasso2d")) %>%
+    layout(yaxis = list(tickmode = "auto", showline = FALSE, gridwidth = 3,
+                        gridcolor = "grey20"),
+           xaxis = list(gridcolor = "grey60"),
+           title = list(xanchor = "left", x = 0,
+                        pad = list(b = 50),
+                        yref = "container", automargin = TRUE),
+           showlegend = TRUE)
+
+  if (is.character(caption)) {
+    plotly_fig <- plotly_fig %>%
+      layout(annotations = list(text = caption,
+                                xref = "paper", yref = "paper",
+                                xanchor = "right", yanchor = "top",
+                                x = 1, y = 1,
+                                showarrow = FALSE,
+                                bgcolor = "white",
+                                font = list(size = 10)))
+  }
+
+  return(plotly_fig)
+}
+
 
 convert_to_Plotly <<- function(ggplot_p,
                                source = "ggplot_source",
@@ -147,55 +168,9 @@ convert_to_Plotly <<- function(ggplot_p,
 
   plot_ly <- ggplotly(ggplot_p, source = source, tooltip = tooltip) %>%
     layout(title = list(text = paste0(ggplot_p$labels$title, "<br>",
-                                      "<sup>", ggplot_p$labels$subtitle, "</sup>")),
-           annotations = list(text = ggplot_p$labels$caption,
-                              xref = "paper", yref = "paper",
-                              xanchor="right", yanchor="top",
-                              x = 1, y = 1,
-                              showarrow = FALSE,
-                              bgcolor = "white",
-                              font = list(size = 10)),
-           yaxis = list(tickmode = "auto", showline = FALSE, gridwidth = 3,
-                        gridcolor = "grey20"),
-           xaxis = list(gridcolor = "grey60"),
-           showlegend = TRUE) %>%
-    config(displaylogo = FALSE,
-           modeBarButtonsToRemove = c("select2d", "lasso2d"))
+                                      "<sup>", ggplot_p$labels$subtitle, "</sup>")))
+  plot_ly <- milquant_plotly_layout(plot_ly, caption = ggplot_p$labels$caption)
   return(plot_ly)
 }
 
-idf_uid_query <<- function(login_connection, uids) {
-  message("Started the query...")
-  query <- paste('{ "selector": { "_id": { "$in": [',
-                 paste0('"', uids, '"', collapse = ", "),
-                 '] } }}',
-                 sep = "")
-  proj_client <- idaifieldR:::proj_idf_client(login_connection,
-                                 include = "query")
-  message("Loading...")
 
-  response <- proj_client$post(body = query)
-  response <- idaifieldR:::response_to_list(response)
-  message("Done.\nProcessing...")
-
-  result <- lapply(response$docs,
-                   function(x) list("id" = x$resource$id, "doc" = x))
-
-
-  result <- idaifieldR:::name_docs_list(result)
-  result <- idaifieldR:::type_to_category(result)
-
-  new_names <- lapply(result, function(x) x$identifier)
-  new_names <- unlist(new_names)
-  names(result) <- new_names
-
-  message("Done.")
-
-  attr(result, "connection") <- login_connection
-  attr(result, "projectname") <- login_connection$project
-  message("Getting config at this point:")
-  attr(result, "config") <- get_configuration(login_connection)
-
-  result <- structure(result, class = "idaifield_docs")
-  return(result)
-}